first RAG production app/pyproject.toml
[project]
name = "first-rag-production-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.9",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.10",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "streamlit>=1.52.0",
    "uvicorn>=0.38.0",
]
